requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
requests==2.31.0
PyPDF2==3.0.1
pinecone==5.0.1
google-generativeai==0.3.2
supabase==2.3.4
sentence-transformers==2.2.2
huggingface_hub==0.16.4
transformers==4.21.3
torch==2.1.0
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.0.3
aiohttp==3.9.1
httpx==0.25.2
gotrue==1.3.0
postgrest==0.13.2
realtime==1.0.6
websockets==11.0.3
pydantic==2.5.0